Fix all ColorVariable warnings
While I’m at it I’ve configured most of the Flat UI theme colors we use
in variables with the same matching name from the theme. Some are only
used once but figured it’d be good to just configure them by name so
its easier to find which ones we are using if someone else is curious
or wants to contribute.